Move file-related functions and classes from session to _utils
- Rename _TruncatableWriter to _SeekableWriter
