+++ dill/_utils.py
@@ -11,22 +11,30 @@ avoid circular import problems.
 
 from __future__ import annotations
 
-__all__ = ['FilterRules', 'Filter', 'RuleType', 'size_filter', 'EXCLUDE', 'INCLUDE']
+__all__ = [
+    'Filter', 'FilterFunction', 'FilterRules', 'FilterSet', 'NamedObject',
+    'Rule', 'RuleType', 'size_filter', 'EXCLUDE', 'INCLUDE',
+]
 
+import contextlib
+import io
 import math
 import re
 import warnings
 from dataclasses import dataclass, field, fields
-from collections.abc import MutableSet
+from collections import abc
+from contextlib import suppress
 from enum import Enum
 from functools import partialmethod
 from itertools import chain, filterfalse
+
+from dill import _dill  # _dill is not completely loaded
+
+# Type hints.
 from typing import (
     Any, Callable, Dict, Iterable, Iterator,
     Optional, Pattern, Set, Tuple, Union,
 )
-
-from dill import _dill
 
 def _format_bytes_size(size: Union[int, float]) -> Tuple[int, str]:
     """Return bytes size text representation in human-redable form."""
@@ -40,7 +48,87 @@ def _format_bytes_size(size: Union[int, float]) -> Tuple[int, str]:
     return size, unit
 
 
-## Namespace filtering. ##
+## File-related utilities ##
+
+class _PeekableReader(contextlib.AbstractContextManager):
+    """lightweight readable stream wrapper that implements peek()"""
+    def __init__(self, stream, closing=True):
+        self.stream = stream
+        self.closing = closing
+    def __exit__(self, *exc_info):
+        if self.closing:
+            self.stream.close()
+    def read(self, n):
+        return self.stream.read(n)
+    def readline(self):
+        return self.stream.readline()
+    def tell(self):
+        return self.stream.tell()
+    def close(self):
+        return self.stream.close()
+    def peek(self, n):
+        stream = self.stream
+        try:
+            if hasattr(stream, 'flush'):
+                stream.flush()
+            position = stream.tell()
+            stream.seek(position)  # assert seek() works before reading
+            chunk = stream.read(n)
+            stream.seek(position)
+            return chunk
+        except (AttributeError, OSError):
+            raise NotImplementedError("stream is not peekable: %r", stream) from None
+
+class _SeekableWriter(io.BytesIO, contextlib.AbstractContextManager):
+    """works as an unlimited buffer, writes to file on close"""
+    def __init__(self, stream, closing=True, *args, **kwds):
+        super().__init__(*args, **kwds)
+        self.stream = stream
+        self.closing = closing
+    def __exit__(self, *exc_info):
+        self.close()
+    def close(self):
+        self.stream.write(self.getvalue())
+        with suppress(AttributeError):
+            self.stream.flush()
+        super().close()
+        if self.closing:
+            self.stream.close()
+
+def _open(file, mode, *, peekable=False, seekable=False):
+    """return a context manager with an opened file-like object"""
+    readonly = ('r' in mode and '+' not in mode)
+    if not readonly and peekable:
+        raise ValueError("the 'peekable' option is invalid for writable files")
+    if readonly and seekable:
+        raise ValueError("the 'seekable' option is invalid for read-only files")
+    should_close = not hasattr(file, 'read' if readonly else 'write')
+    if should_close:
+        file = open(file, mode)
+    # Wrap stream in a helper class if necessary.
+    if peekable and not hasattr(file, 'peek'):
+        # Try our best to return it as an object with a peek() method.
+        if hasattr(file, 'tell') and hasattr(file, 'seek'):
+            file = _PeekableReader(file, closing=should_close)
+        else:
+            try:
+                file = io.BufferedReader(file)
+            except Exception:
+                # It won't be peekable, but will fail gracefully in _identify_module().
+                file = _PeekableReader(file, closing=should_close)
+    elif seekable and (
+        not hasattr(file, 'seek')
+        or not hasattr(file, 'truncate')
+        or (hasattr(file, 'seekable') and not file.seekable())
+    ):
+        file = _SeekableWriter(file, closing=should_close)
+    if should_close or isinstance(file, (_PeekableReader, _SeekableWriter)):
+        return file
+    else:
+        return contextlib.nullcontext(file)
+
+
+## Namespace filtering ##
 
 RuleType = Enum('RuleType', 'EXCLUDE INCLUDE', module=__name__)
 EXCLUDE, INCLUDE = RuleType.EXCLUDE, RuleType.INCLUDE
@@ -73,7 +161,7 @@ def _iter(filters):
         return None
 
 @dataclass
-class FilterSet(MutableSet):
+class FilterSet(abc.MutableSet):
     """A superset of exclude/include filter sets."""
     names: Set[str] = field(default_factory=set)
     regexes: Set[Pattern[str]] = field(default_factory=set)
@@ -396,9 +484,7 @@ class FilterRules:
         return namespace_copy
 
 
-######################
-#  Filter factories  #
-######################
+## Filter factories ##
 
 import collections
 import collections.abc
